✅ test: rename sendGroup to starter and add Agent test
- Rename sendGroup.feature/steps to starter.feature/steps
- Add E2E test for sendAsAgent sidebar refresh
- Both Agent and Group creation now have E2E test coverage
